fixing

+++ movierecommendationsystem-frontend/src/app/favorite-movies/favorite-movies.component.ts
@@ -9,6 +9,7 @@ import { Router } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
 import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 import { General } from '../model/generalInfo';
+import { User } from '../model/user';
 
 @Component({
   selector: 'app-favorite-movies',
@@ -20,7 +21,7 @@ import { General } from '../model/generalInfo';
 export class FavoriteMoviesComponent implements OnInit {
   favorite : General[] = [];
   username : string = '';
-  user : any;
+  user : User | null = null;
   constructor(private userService : UserService,private authService : AuthService,
     private interactionService : InteractionService, private router : Router,
   private dialog : MatDialog){}
@@ -30,10 +31,12 @@ export class FavoriteMoviesComponent implements OnInit {
         this.userService.findByUsename(this.username).subscribe(
           response => {
             this.user = response;
-            this.interactionService.favorite(this.user.userId).subscribe(
-              response => {this.favorite = response;},
-              error => {console.error('error fetching watch list',error);}
-            );
+            if(this.user){
+              this.interactionService.favorite(this.user.userId).subscribe(
+                response => {this.favorite = response;},
+                error => {console.error('error fetching watch list',error);}
+              );
+            }
           },
           error => {console.error('error fetching user',error);}
         );
@@ -47,7 +50,7 @@ export class FavoriteMoviesComponent implements OnInit {
       data : {message : 'Are you sure you want to remove this movie from your favorite?'}});
     dialogRef.afterClosed().subscribe(
       response => {
-        if(response){
+        if(response && this.user){
           this.interactionService.removeFromFavorite(this.user.userId,movieId).subscribe(
             response => {this.ngOnInit();},
             error => {console.error('error deleting movie from favorite',error);}
